Implementation/rag_phase1/requirements.txt
sentence-transformers
faiss-cpu
psycopg2-binary
numpy
scikit-learn
